Remove valid page ID check from the eviction logic
There isn't a valid case for a frame having an invalid page. If a frame is
in use, then the page it contains should be valid. Even if we create a brand
new page and load it into a frame, it's immediately valid because when
we allocate space for it, we'll get a valid page ID.