Implement working internal node split propagation
This works, but is not really correct. There is no case in which an
internal node would split unless a child node had already split. This means that we
always need to account for the new right side sibling of the splitting parent node's
split child. The SplitAndInsert should be setting the original node's rightmost pointer
to the new right sibling of the original child node that has been split.