Add test for promoted key being the largest in the parent node and fixed the logic
If the promoted key being inserted into a parent internal node is the largest
key, then we need to set the rightmost pointer of the parent node to point
to the new right child created as part of the split. Otherwise, we find the
separator key to the right of the new promoted key and set it's pointer to
the new right child.